sim_derive, models: Group generated functions in SerializableModel trait
