+++ Cargo.toml
@@ -35,6 +35,7 @@ serde_yaml = "0.8"
 thiserror = "1.0"
 wasm-bindgen = "0.2.63"
 web-sys = { version = "0.3", features = [ "console" ] }
+sim_derive = { path = "sim_derive" }
 
 # The `console_error_panic_hook` crate provides better debugging of panics by
 # logging them with `console.error`. This is great for development, but requires
