+++ sim_derive/Cargo.toml
@@ -4,3 +4,7 @@ version = "0.7.1"
 
 [lib]
 proc-macro = true
+
+[dependencies]
+syn = "*"
+quote = "*"
