+++ src/models/generator.rs
@@ -2,13 +2,15 @@ use std::f64::INFINITY;
 
 use serde::{Deserialize, Serialize};
 
-use super::model_trait::AsModel;
+use super::model_trait::{AsModel, SerializableModel};
 use super::ModelMessage;
 use crate::input_modeling::random_variable::ContinuousRandomVariable;
 use crate::input_modeling::Thinning;
 use crate::simulator::Services;
 use crate::utils::error::SimulationError;
 use crate::utils::{populate_history_port, populate_snapshot_port};
+
+use sim_derive::SerializableModel;
 
 /// The generator produces jobs based on a configured interarrival
 /// distribution. A normalized thinning function is used to enable
@@ -18,7 +20,7 @@ use crate::utils::{populate_history_port, populate_snapshot_port};
 /// produce jobs through perpetuity, and the generator does not receive
 /// messages or otherwise change behavior throughout a simulation (except
 /// through the thinning function).
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, SerializableModel)]
 #[serde(rename_all = "camelCase")]
 pub struct Generator {
     // Time between job generations
@@ -123,13 +125,6 @@ impl Generator {
             history: Default::default(),
         }
     }
-    
-    pub fn from_value(value: serde_yaml::Value) -> Option<Box<dyn AsModel>> {
-        match serde_yaml::from_value::<Self>(value) {
-            Ok(model) => Some(Box::new(model)),
-            Err(_) => None
-        }
-    }
 
     fn need_snapshot_metrics(&self) -> bool {
         self.ports_in.snapshot.is_some() && self.ports_out.snapshot.is_some()
@@ -143,14 +138,6 @@ impl Generator {
 }
 
 impl AsModel for Generator {
-    fn get_type(&self) -> &'static str {
-        "Generator"
-    }
-
-    fn serialize(&self) -> serde_yaml::Value {
-        serde_yaml::to_value(self).unwrap_or(serde_yaml::Value::Null)
-    }
-    
     fn status(&self) -> String {
         format!["Generating {}s", self.ports_out.job]
     }
